Update vpc/variables.tf
To fix an issue where the GKE cluster was running out of IP addresses during provisioning. Note that the IP addresses entered in the variables file get changed during provisioning. Not sure why this is the case. The subnet mask +=4, and the IP changes slightly as well. The default values have been adjusted accordling, e.g. /12 in order to end up with /16.